change date format in model,add tables name in all models, add api for getting all products,add isvalidateuser api

+++ database/migrations/2020_05_02_114444_create_prices_table.php
@@ -17,7 +17,7 @@ class CreatePricesTable extends Migration
             $table->id();
             $table->foreignId('user_id');
             $table->foreignId('stock_id');
-            $table->integer('price');
+            $table->double('price', 10, 2);
             $table->timestamps();
         });
     }
